# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de base en laravel",
        "description": "GabomaPay",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://gabomapay.com/api"
        }
    ],
    "paths": {
        "/admin/admins/list": {
            "get": {
                "tags": [
                    "ADMINISTRATION__ADMINISTRATEUR"
                ],
                "summary": "Liste des administrateurs",
                "operationId": "70d159651ab1842498461cf31a7343ba",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/admin/admins/store": {
            "post": {
                "tags": [
                    "ADMINISTRATION__ADMINISTRATEUR"
                ],
                "summary": "Ajouter un administrateur",
                "operationId": "82cb61d613d89b30f79428737c43ee6c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mulamba"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Enock"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "bgi@gmail.com"
                                    },
                                    "sexe": {
                                        "type": "string",
                                        "example": "M"
                                    },
                                    "country": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "admin_level": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/admins/show/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__ADMINISTRATEUR"
                ],
                "summary": "Renvoie les informations d'un administrateur par son ID",
                "operationId": "31681fa15426ac49b7e17ffd74f4351f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/admins/update": {
            "post": {
                "tags": [
                    "ADMINISTRATION__ADMINISTRATEUR"
                ],
                "summary": "Editer les informations d'un administrateur",
                "operationId": "b8366dfcf03e7f01de3503e2f111d91d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Mulamba"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Enock"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "bgi@gmail.com"
                                    },
                                    "admin_level": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/admins/activate/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__ADMINISTRATEUR"
                ],
                "summary": "Active ou désactive l'état d'un administrateur par son ID",
                "operationId": "6402cc5b0e938b02ecdcbea9cfa05f43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            },
            "0": []
        }
    }
}